// src/types/vehiclelistreports.ts
export type DeviceAttach = {
  true: boolean;
  IgnitionStatus: boolean;
  IsActive: boolean;
  Label1: string;
  clientId: string;
  currentDriverName: string;
  driverStatus: boolean;
  id: string;
  vehicleMake: string;
  vehicleModel: string;
  vehicleType: string;
  vehicleNo: string;
  vehicleReg: string;
  vehicleskipstep: number;
  deviceIMEI: string;
  deviceId: string;
  dualCam: boolean;
  immobilising: boolean;
  camStatus: string;
  immStatus: string;
  __v: number;
  _id: string;
  fuelTankCapacity: number 

};